requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
feedparser>=6.0.0
cachetools>=5.3.0
